fix: align iOS workspace title glass
