+++ Packages/iOS/CmuxMobileShellUI/Sources/CmuxMobileShellUI/WorkspaceTitleMenu.swift
@@ -28,7 +28,7 @@ struct WorkspaceTitleMenu<Label: View, MenuContent: View>: View {
     }
 
     private var fittedLabel: some View {
-        let cap = MobileNavTitleWidth(
+        let cap = MobileLeadingToolbarTitleWidth(
             contentWidth: contentWidth,
             hasBackButton: hasBackButton,
             hasTrailingCluster: hasTrailingCluster,
@@ -37,7 +37,7 @@ struct WorkspaceTitleMenu<Label: View, MenuContent: View>: View {
 
         return label()
             .frame(
-                minWidth: min(MobileNavTitleWidth.floor, cap),
+                minWidth: min(MobileLeadingToolbarTitleWidth.floor, cap),
                 maxWidth: cap,
                 alignment: .leading
             )
